chore: Change size_sqm column type to float in areas table
Updated the migration for the areas table to use a float type for the size_sqm column instead of decimal. This may improve compatibility with calculations and storage of area sizes.

+++ database/migrations/2025_08_19_055106_create_areas_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('name');
             $table->enum('type', AreaTypeEnum::values());
             $table->text('description')->nullable();
-            $table->decimal('size_sqm', 8, 2)->nullable();
+            $table->float('size_sqm', 8)->nullable();
             $table->json('coordinates')->nullable(); // For positioning within garden
             $table->json('dimensions')->nullable(); // Length, width, height if applicable
             $table->string('color', 7)->nullable(); // Hex color for visualization
